Update README.md
This note and project description highlight the academic structure and practical application of the skills taught in the course, reflecting the real-world application of Android development principles in collaboration with Combitech. The login feature serves as a crucial gateway to the application, ensuring secure user access and interaction with Navelink services.